GOV-167 Renaming incorrectly named files back into Payments section
As per the issue at GitBook and now resolved, and "out of date" merge request on the GitBook system used to be able to apply changes that made no sense. They have now rolled out a new system and UI that does the equivalent of a rebase on the merge request, includign a 3 way merge if necessary. Should stop us falling into same trap again.